\name{myMapApp-package}
\alias{myMapApp-package}
\alias{myMapApp}
\docType{package}
\title{
\packageTitle{myMapApp}
}
\description{
\packageDescription{myMapApp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myMapApp}
\packageIndices{myMapApp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{myMapApp}

Maintainer: \packageMaintainer{myMapApp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
