% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMapApp.R
\docType{package}
\name{myMapApp}
\alias{myMapApp}
\alias{myMapApp-package}
\title{myMapApp}
\description{
A package which implements Shiny package to make an 
   app for myMap package. This app provides an interface
   for user to implement the myMap interactively.
}
\seealso{
[https://shiny.rstudio.com/]
}
\author{
Ali Etminan, Mohsen Pirmoradiyan
}
