% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample()
}
\value{
Shiny app to run myMap.
}
\description{
This function runs the shiny myMapApp. The app shows the myMap methods
    output interactively.
}
\details{
\code{runExample()} Run the shiny app related to myMap package.
}
